Rename Pre-Main class
