+++ opentelemetry-javaagent/opentelemetry-javaagent.gradle
@@ -17,9 +17,9 @@ configurations {
 jar {
   manifest {
     attributes(
-      "Main-Class": "io.opentelemetry.auto.bootstrap.AgentBootstrap",
-      "Agent-Class": "io.opentelemetry.auto.bootstrap.AgentBootstrap",
-      "Premain-Class": "io.opentelemetry.auto.bootstrap.AgentBootstrap",
+      "Main-Class": "io.opentelemetry.javaagent.OpenTelemetryAgent",
+      "Agent-Class": "io.opentelemetry.javaagent.OpenTelemetryAgent",
+      "Premain-Class": "io.opentelemetry.javaagent.OpenTelemetryAgent",
       "Can-Redefine-Classes": true,
       "Can-Retransform-Classes": true,
     )
